# Conflicts:
#	app/src/main/java/com/arvid/dtuguide/Main2Activity.java
#	app/src/main/java/com/arvid/dtuguide/data/MARKTYPE.java
#	app/src/main/java/com/arvid/dtuguide/navigation/NavigationController.java
#	app/src/main/res/mipmap-hdpi/ic_launcher.png
#	app/src/main/res/mipmap-hdpi/ic_launcher_round.png
#	app/src/main/res/mipmap-mdpi/ic_launcher.png
#	app/src/main/res/mipmap-mdpi/ic_launcher_round.png
#	app/src/main/res/mipmap-xhdpi/ic_launcher.png
#	app/src/main/res/mipmap-xhdpi/ic_launcher_round.png
#	app/src/main/res/mipmap-xxhdpi/ic_launcher.png
#	app/src/main/res/mipmap-xxhdpi/ic_launcher_round.png
#	app/src/main/res/mipmap-xxxhdpi/ic_launcher.png
#	app/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/arvid/dtuguide/SettingsActivity.java
@@ -107,7 +107,7 @@ public class SettingsActivity extends AppCompatActivity implements View.OnClickL
                 builder.setTitle(R.string.dialog_zoom_title);
                 builder.setPositiveButton(R.string.save, new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int id) {
-                        Settings.getInstance(getApplicationContext()).setZoom(Float.parseFloat((String)zoomLevelTV.getText()));
+                        Settings.getInstance(getApplicationContext()).setZoom(Integer.parseInt((String)zoomLevelTV.getText()));
                     }
                 });
                 builder.setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
